requirements.txt
ansicolors
numpy
ply>=3.11
PyYAML>=3.13
